added resource builder
The resourcemanager is reusing allot of code so making a resource builder was another priority before adding shadows and other passes.

Materials are currently stored inside the meshes vector which makes it harder to store it inside a data* variable.

I still need to add texture building and sampler building